[21.11.08] # Board App

+++ src/main/java/com/spring/mvc/board/repository/BoardRepository.java
@@ -10,13 +10,13 @@ public interface BoardRepository {
     List<Board> getArticles();
 
     //게시물 상세 조회
-    Board getContent(int boardNo);
+    Board getContent(Long boardNo);
 
     //게시물 등록
     boolean insert(Board board);
 
     //게시물 삭제
-    boolean delete(int boardNo);
+    boolean delete(Long boardNo);
 
     //게시물 수정
     boolean update(Board board);
